// src/sanity/sanity.types.ts
export type Slug = {
	_type: "slug";
	current?: string;
	source?: string;
};

export type PortfolioProjects = {
	_id: string;
	_type: "portfolio-projects";
	_createdAt: string;
	_updatedAt: string;
	_rev: string;
	title?: string;
	slug?: Slug;
	featured?: boolean;
	mainImage?: {
		asset?: {
			_ref: string;
			_type: "reference";
			_weak?: boolean;
			[internalGroqTypeReferenceTo]?: "sanity.imageAsset";
		};
		media?: unknown;
		hotspot?: SanityImageHotspot;
		crop?: SanityImageCrop;
		alt?: string;
		caption?: string;
		_type: "image";
	};
	coverVideo?: Youtube;
	trailerVideo?: Youtube;
	localIntroVideo?: string;
	categories?: Array<{
		_ref: string;
		_type: "reference";
		_weak?: boolean;
		_key: string;
		[internalGroqTypeReferenceTo]?: "project-category";
	}>;
	tags?: Array<string>;
	publishedAt?: string;
	description?: string;
	body?: Array<
		| {
				children?: Array<{
					marks?: Array<string>;
					text?: string;
					_type: "span";
					_key: string;
				}>;
				style?: "normal" | "h1" | "h2" | "h3" | "h4" | "blockquote";
				listItem?: "bullet";
				markDefs?: Array<{
					href?: string;
					_type: "link";
					_key: string;
				}>;
				level?: number;
				_type: "block";
				_key: string;
		  }
		| {
				asset?: {
					_ref: string;
					_type: "reference";
					_weak?: boolean;
					[internalGroqTypeReferenceTo]?: "sanity.imageAsset";
				};
				media?: unknown;
				hotspot?: SanityImageHotspot;
				crop?: SanityImageCrop;
				alt?: string;
				caption?: string;
				_type: "image";
				_key: string;
		  }
		| ({
				_key: string;
		  } & Youtube)
	>;
};

export type SanityImageCrop = {
	_type: "sanity.imageCrop";
	top?: number;
	bottom?: number;
	left?: number;
	right?: number;
};

export type SanityImageHotspot = {
	_type: "sanity.imageHotspot";
	x?: number;
	y?: number;
	height?: number;
	width?: number;
};

export type Youtube = {
	_type: "youtube";
	url?: string;
};
// ...
export declare const internalGroqTypeReferenceTo: unique symbol;